+++ html/sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'e3dc-control-v5-4-2a-static';
+const CACHE_NAME = 'e3dc-control-v5-4-2b-static';
 const ASSETS = [
   'assets/vendor/bootstrap/css/bootstrap.min.css',
   'assets/vendor/fontawesome/css/all.min.css',